Fork cleanup (#3)
Some fork cleanup since we've pulled this in.  The readme should now better credit the original author on the source fork, [chrnorm/deployment-status](https://github.com/chrnorm/deployment-status).  The examples in the readme now only reference this action, as this is all we're forking.

The removed workflows either work on a schedule (`codeql-analysis.yml`) or require `chrnorm/deployment-action` to run (`test.yml`).

I considered removing [`check-dist.yml`](https://github.com/Mortgage-Research-Center/deployment-status/blob/main/.github/workflows/check-dist.yml), but that seems to be important to the Action we want from this repo.